coms/requirements.txt
numpy==1.21.1
netifaces==0.11.0
pyyaml==6.0
